pyproject.toml
[project]
name = "chemvision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "datasets>=3.6.0",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "tqdm>=4.67.1",
]
